Work around google's buggy EDIT URL business in a new way
Our old workaround for getting approval edit URLs was to just submit the form then take the most recent item from the form. That stopped working reliably in March, 2018, for reasons I don't understand.

A new workaround is to:
(A) submit the form
(B) use the "ID" of the submitted form to get ANOTHER FormResponse object
(C) Use the getEditResponseUrl off of THAT object (which obviously should be the same as the first object, but isn't).

This is working for now. Once again, this is not satisfactory... but it works.

See: https://issuetracker.google.com/issues/36762422